+++ src/main/java/com/_32bit/project/cashier_system/DTO/offer/OfferInfoResponse.java
@@ -15,6 +15,6 @@ public class OfferInfoResponse {
     private String insertionDate;
     private String startDate;
     private String endDate;
-    private Integer getCount;
-    private Integer payFor;
+    private Double getCount;
+    private Double payFor;
 }
